Move ProcessorFactory to processor.Factory
