+++ internal/collector/processor/nodebatcher/factory.go
@@ -18,12 +18,11 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/open-telemetry/opentelemetry-service/consumer"
-	"github.com/open-telemetry/opentelemetry-service/factories"
 	"github.com/open-telemetry/opentelemetry-service/models"
 	"github.com/open-telemetry/opentelemetry-service/processor"
 )
 
-var _ = factories.RegisterProcessorFactory(&processorFactory{})
+var _ = processor.RegisterProcessorFactory(&processorFactory{})
 
 const (
 	// The value of "type" key in configuration.
